# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PupiBot Agent Runner API",
    "description": "API para gestión y ejecución de operaciones por agentes",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/agent/execute": {
      "post": {
        "summary": "Execute Action",
        "operationId": "execute_action_v1_agent_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/operations/{operation_id}": {
      "get": {
        "summary": "Get Operation Status",
        "operationId": "get_operation_status_v1_agent_operations__operation_id__get",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/operations": {
      "post": {
        "summary": "Create a new operation",
        "description": "Creates a new operation with the provided action and payload.",
        "operationId": "create_operation_operations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/operations/{operation_id}/run": {
      "post": {
        "summary": "Run an operation",
        "description": "Executes the specified operation and updates its status.",
        "operationId": "run_operation_operations__operation_id__run_post",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExecuteRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "action",
          "payload",
          "idempotency_key"
        ],
        "title": "ExecuteRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Operation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "ID único de la operación"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Tipo de acción que ejecuta la operación"
          },
          "payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload",
            "description": "Datos asociados a la operación"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Estado actual de la operación"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "Resultado de la operación"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action"
        ],
        "title": "Operation"
      },
      "OperationIn": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "action",
          "payload"
        ],
        "title": "OperationIn"
      },
      "OperationResponse": {
        "properties": {
          "operation_id": {
            "type": "integer",
            "title": "Operation Id"
          }
        },
        "type": "object",
        "required": [
          "operation_id"
        ],
        "title": "OperationResponse"
      },
      "OperationStatus": {
        "properties": {
          "operation_id": {
            "type": "integer",
            "title": "Operation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "status"
        ],
        "title": "OperationStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}